\name{meteorogram-package}
\alias{meteorogram-package}
\alias{meteorogram}
\docType{package}
\title{
\packageTitle{meteorogram}
}
\description{
\packageDescription{meteorogram}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{meteorogram}
\packageIndices{meteorogram}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{meteorogram}

Maintainer: \packageMaintainer{meteorogram}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
